requirements.txt
ccxt
pandas
python-dotenv
python-telegram-bot==13.15